+++ processing/subsea_cable_processing_provider.py
@@ -23,6 +23,7 @@ from .import_ship_outline_algorithm import ImportShipOutlineAlgorithm
 from .place_ship_outlines_algorithm import PlaceShipOutlinesAlgorithm
 from .plot_line_segments_from_table_algorithm import PlotLineSegmentsFromTableAlgorithm
 from .translate_kp_from_rpl_to_rpl_algorithm import TranslateKPFromRPLToRPLAlgorithm
+from .rpl_route_comparison_algorithm import RPLRouteComparisonAlgorithm
 
 
 class SubseaCableProcessingProvider(QgsProcessingProvider):
@@ -59,6 +60,8 @@ class SubseaCableProcessingProvider(QgsProcessingProvider):
         self.addAlgorithm(PlotLineSegmentsFromTableAlgorithm())
         print('Registering TranslateKPFromRPLToRPLAlgorithm...')
         self.addAlgorithm(TranslateKPFromRPLToRPLAlgorithm())
+        print('Registering RPLRouteComparisonAlgorithm...')
+        self.addAlgorithm(RPLRouteComparisonAlgorithm())
 
     def id(self):
         """